Ref : Extracted Navbar into smaller sub-components #2085
